Allow sent tags to be configured
I prefer to add an inbox tag to messages I send. That way I can see when I haven't got an answer, etc. This adds an inbox tag to the sent tag of sent emails. Ideally of course, this shouldn't be hardcoded but, like "excluded_tags" in the config file.