ti: use state-less queries, re-querying when loading more threads.
keeping notmuch query objects alive for a long while causes issues since
they are quite fragile to outside (or inside database changes). all
internal changes to the database (including polling with notmuch lastmod
functionality) should be caught.

the downside is that it takes longer to load more threads, hopefully
this is fixed when notmuch_threads_move_to_next is fixed and actually
moves the iterator.