actions, mt, tv: document and clean up emitted signals
describes and cleans up emitted signals. Notmuch* objects do not take
care of themselves. Message* objects do, users of those should rely on
the relayed 'message-changed' signal to make sure that the Message*
object is up-to-date.

There will probably be a 'message-thread-changed' signal in the future
when we need it.