poll, astroid: add --start-polling and --stop-polling to indicate external polling
See #64, #275. This adds the --start-polling and --stop-polling flags to
astroid which indicates that there is an external polling operation
going on. This starts the spinner, prevents internal polling,
and refreshes any threads that have been changed since --start-polling
when --stop-polling is called.

It is the users responsibility to make sure to call --stop-polling after
having called --start-polling, even if the polling operation fails.