compiler: should use __has_cpp_attribute and check for availability
The build failed when compiling src/utils/gmime/gmime-filter-html-bq.c.
It complains about a missing ')' after 'clang' in line 5 of compiler.h.
This is because clang::fallthrough is a scoped attribute, but the used
macro __has_attribute will only accept single identifiers under clang.
Hence, its sibling __has_cpp_attribute must be used. Then, when using
the LLVM clang that ships with Xcode, the __has_cpp_attribute is not
defined at all when compiling gmime-filter-html-bq.c. This is because it
is a C file, and the LLVM clang that ships with Xcode hides the C++
features (including the feature checking macros) when operating in C
mode. Thus we must safeguard against the __has_* macros not being
defined. For further details, see

https://isocpp.org/std/standing-documents/sd-6-sg10-feature-test-recommendations

and

https://clang.llvm.org/docs/LanguageExtensions.html#feature-checking-macros

Signed-off-by: Gaute Hope <eg@gaute.vetsj.com>