cmake: add config switch for embedded editor
New switch to disable building the embedded editor. The editor is built
by default.

On macOS, the build with the embedded editor breaks on a missing
gtkmm/socket.h. This is because currently the macOS build uses an old
webkitgtk version. Hence, if a macOS build is detected, the embedded
editor is forcibly decativated, even if it was explicitly activated on
the command line. Once the new webkitgtk version gets integarted, we can
remove the macOS build check in lines 204 through 207.

This fixes c-alpha/astroid#5