Use Hard-Coded Version Number for Release Builds
This removes the need for on-line connection when building a release
version. For debug builds, the version number is retrieved from github,
and the commit hash is appended.

Since the update to macOS 10.13, I need to specify include_directories().
No idea why...

In target_link_libraries(), switching to the _LDFLAGS macros, since they
also contain the path specs (-L options). This is also caused by the
macOS 10.13 update.