Enhance Message ID Left Side Generation
This makes the mail.message_id_user optional. If it is not set (empty
string), nothing is inserted immediately before the "@". To retain the
previous factory default, config.cc has also been updated to set the
mail.message_id_user to "astroid" when a factory default config file is
written.

The gerenation process for the random part has been updated to no longer
use the time, but instead generate 40 random characters from the full set
of 81 allowed charaters for the left side of the message id as per
RFC 5322. This provides 253 bit of entropy in the left side string.
Also, the random number generation now uses arc4random. Picking "the best"
random number generator is practically impossible [1][2][3]. With
arc4random I have made the least worse experiences across systems.
[1] https://security.stackexchange.com/questions/85601/is-arc4random-secure-enough
[2] https://www.2uo.de/myths-about-urandom/
[3] https://calomel.org/entropy_random_number_generators.html

As RFC5322 message ids are not necessarily legal file names, the
generation of the tmp file name for external editors also had to be
modified. Now, the tmp file name is generated independently of the
message id. To allow external editors to recognise email message files
(e.g. to switch to an email editing mode of sorts), a new configuration
variable "editor.external_tmpfile_suffix" was introduced, with a default
value of "@msg.draft". This will be used as the suffix of all tmp files
handed off to external editors. Consequently, the CONFIG_VERSION had to
be bumped to 10.