Enhance Message ID Right Side Generation
The config variable "mail.message_id_fqdn" can now also contain a "%1"
parameter, which will be replaced with the domain name of the email
address contained in the from header of the message being composed. When
the from address of the message being composed is changed (when modified
in an extrenal editor, or via the drop-down menu after editing), the
right-hand side (i.e. the stuff after the "@") of the message id gets
updated with the new domain name.

This is useful whe a user has several email accounts for sending, and
wishes to always set the host and domain name part of the message id to
the domain name of the address he is sending the particular message from.