catch more shared lib naming conventions
It seems that on some platforms, a ".so" is inserted before the micro version.
Patterns catched so far:
libgmime.1.2.3.so
libgmime-1.2.3.so
libgmime.so.1.2.3
New patterns added with this commit:
libgmime.1.2.so.3
libgmime-1.2.so.3

Also, system libraries are inlcuded in the search, sine gmime & friends get installed in '/usr/lib` on some platforms.