Copy Tags to New Draft Before Deleting Old Draft
The tags were copied _after_ the old draft message was deleted. That this worked was pure coincidence. A copy of the tags struct would have to be made, but
a notmuch_tags_t can not be copied, because the data type is opaque, and the compiler cannot know its size.

To be on the safe side, we are thus now copying the tags when the old draft is still around. This reordering of steps also requires that the new draft be saved in a file with a different file name. Before, we reused the same file name, because the old draft was completely deleted before the new draft was created (and hence it was afe to re-use the same file name).