Refactor Message ID Regeneration One Level Up
It turns out that EditMessage::read_edited_message() is always called, even if only the sender opo-up in the UI is operated. Hence refactored the code to update the message id from EditMessage::set_from (Account * a) to EditMessage::read_edited_message() (which itself calls set_from()). This has the advantagem that we can operate directly on the ComposeMessage which is being handed through the editing process, instead of creating yet another one in set_from().

TODO: After the message id has been updated for a draft message which was alone on a thread (i.e. not a reply or forward), notmuch issues a new thread id. In this case all UI thread list views that show queries involving the "draft" tag need to be refreshed.