Prevent ComposeMessage From Being Delete Topo Early In EditMessage::read_edited_message ()
There was a race condition in EditMessage::read_edited_message(). That method spawns an OnMessageAction when the message id needs to be updated. A ComposeMessage, which is created via new() as a local instance in EditMessage::read_edited_message(), is passed into the lambda and used there. Right after the lambda was queued, the local ComposeMessage was deleted again near the end of EditMessage::read_edited_message(). This lead to a high probability of it having been deleted before the lambda of the OnMessageAction was executed by the ActionManager. As a result, the lambda code ran with a reference to freed memeory.

By putting the deletion of the local ComposeMessage in its own OnMessageAction, it is ensured that it runs after the OnMessageAction to update the message id (since it sits further down in the same queue). Consequently, the lambda code to update the message id is guaranteed to have a valid ComposeMessage available, and the local ComposeMessage instance from EditMessage::read_edited_message() is deleted as soon as it is no longer needed.