add syslog backend and make logging sinks and level configurable
remove file log and add syslog. the multi-process design we have now
would require careful handling when output to one file.

syslog and stdout config options for the log are created. as well as one
config option for setting the log-level (fixes #419). the corresponding
log-level option is removed as executable argument. a --log-stdout
argument is added. and --disable-log still overrides.