fix #550: use srcdoc attribute when setting message content in iframe
using the src attribute causes a request that needs to be allowed on the
main GUI thread on new webkit. even when setting the iframe src async it
needs to run on the extension GUI thread in order to be able to
manipulate the DOM tree. thus subsequent calls to the webext, also
necessarily running on the extension GUI thread, blocks .. and the main
GUI blocks while waiting for these to return. this causes a deadlock
while the webext waits for the request to be handled on the main GUI
thread.