use correct size of html part in multipart messages
fixes #585

The html rendered from markdown is stored in the `ustring _html`.  The
function `g_mime_stream_mem_new_with_buffer()` requires a `char *` as
first argument and its length as second. The correct length of the
`(char *) _html.c_str()` is `_html.bytes()` and not `_html.size()` which
is the number of UTF-8 characters.